movies/views.py
from rest_framework import viewsets
from .models import Movie
from .serializers import MovieSerializer  

class MovieViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = Movie.objects.all().order_by('-created_at')
    serializer_class = MovieSerializer
